Create DOMInputElement class
