+++ src/DOMInputElement.ts
@@ -1,0 +1,9 @@
+export class DOMInputElement {
+  constructor(private id: string) {}
+
+  element = <HTMLInputElement>document.getElementById(this.id);
+
+  get value() {
+    return this.element.value;
+  }
+}
